requirements.txt
# Core Deep Learning
tensorflow==2.17.1
keras==3.5.0
numpy==1.26.4

# Data Processing
scikit-learn==1.2.2
pandas>=2.0.0

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# UI
customtkinter==5.2.2